Fix : 로그인을 위해 refreshToken에 LOGIN 넣어놓은 상테에서는 통신 request 보내지 않도록 null 처리

+++ app/src/main/java/com/depromeet/housekeeper/data/utils/AuthAuthenticator.kt
@@ -19,6 +19,10 @@ class AuthAuthenticator @Inject constructor(
             tokenManager.getRefreshToken().first()
         }
         Timber.d("refreshToken = ${refreshToken}")
+        if (refreshToken == "LOGIN") {
+            response.close()
+            return null
+        }
         if (refreshToken == response.request.header(AUTHORIZATION)) {
             Timber.d("refresh token 이미 보냈었음")
             runBlocking { tokenManager.deleteTokens() }
